cbit/backend/requirements.txt
flask
flask-cors
openai-whisper
tqdm
ffmpeg-python
python-multipart
werkzeug